feat: decorator to simplify view-based object permission checks
https://new-hippo.atlassian.net/browse/HC-134

Neil pointed out some important limitations with the existing approach
using the built-in Guardian decorators. Namely:
- order matters and getting the order wrong can leak implementation
  details
- the decorator is making redundant db queries that we're making in the
  body of the view (this can be confusing and is inconsistent)

This approach introduces a single decorator that does the same
permission checks, but also passes the authorized objects to the view
function.

This includes an example on an existing view -- if this is merged, a
future PR will make this same change across our other permissioned
views.