fix: form builder renders
https://new-hippo.atlassian.net/browse/HC-317

After the datadog changes, environment vars were moved to an object
called `datadog_vars`. Form builder reads from `environment` but it's
fallback was not parseable so it was failing.

This change pulls `environment` from `datadog_vars` and makes the
fallback value parseable so if this happens again, form builder at least
loads.